backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
